backend/sample_agents/personalization_agent.py
#!/usr/bin/env python3
"""
Personalization AI Agent - Git Versioned
This agent generates personalized website variants based on user behavior
"""

import json
from typing import Dict, Any
import anthropic
from openai import OpenAI

class PersonalizationAgent:
    """AI Agent for website personalization with Git versioning"""
    
    def __init__(self, version: str = "1.0.0"):
        self.version = version
        self.claude = anthropic.Anthropic()
        self.morph = OpenAI(base_url="https://api.morphllm.com/v1")
    
    def generate_variant(self, user_data: Dict[str, Any]) -> Dict[str, Any]:
        """Generate personalized website variant"""
        
        # AI logic for personalization
        variant = self._analyze_user_behavior(user_data)
        code = self._generate_component_code(variant)
        
        return {
            "success": True,
            "variant": variant,
            "generated_code": code,
            "agent_version": self.version,
            "timestamp": "2024-01-15T10:30:00Z"
        }
    
    def _analyze_user_behavior(self, user_data: Dict[str, Any]) -> str:
        """Analyze user behavior to determine variant type"""
        # Implementation would use Claude for analysis
        return "pricing_focused"
    
    def _generate_component_code(self, variant_type: str) -> str:
        """Generate React component code using Morph"""
        # Implementation would use Morph for code generation
        return "// Generated React component..."

if __name__ == "__main__":
    agent = PersonalizationAgent()
    print(f"Personalization Agent v{agent.version} ready!")
